Added the stuff for toggle brightness

+++ VortexEngine/src/Leds/Leds.cpp
@@ -25,7 +25,7 @@
 // array of led color values
 RGBColor Leds::m_ledColors[LED_COUNT] = { RGB_OFF };
 // global brightness
-uint32_t Leds::m_brightness = DEFAULT_BRIGHTNESS;
+uint8_t Leds::m_brightness = DEFAULT_BRIGHTNESS;
 
 // Output PORT register
 volatile uint8_t *Leds::m_port = nullptr;
